Separate TLObject
