+++ tl_generator/main.ts
@@ -5,20 +5,7 @@ import { revampId, revampType } from "./utilities.ts";
 const tlContent = Deno.readTextFileSync("tl/api.tl");
 
 const { constructors } = parse(tlContent);
-let code = `// deno-lint-ignore-file no-explicit-any
-
-const id = Symbol("id");
-
-const params = Symbol("params");
-
-export abstract class TLObject {  
-  protected abstract get [id](): number | symbol;
-  protected abstract get [params](): Params | symbol;
-}
-
-type MaybeInArray<T> = T | [T];
-
-type Params = [string | number | null | bigint | boolean | Uint8Array | Array<any> | TLObject, MaybeInArray<typeof TLObject | typeof Uint8Array | string>][];
+let code = `import { id, params, TLObject, Params } from "./tl_object.ts";
 
 export abstract class Constructor extends TLObject {
 }
